+++ packages/core/src/markView/CoreMarkView.ts
@@ -2,7 +2,7 @@ import { isElementLike } from '@ocavue/utils'
 import type { Mark } from 'prosemirror-model'
 import type { EditorView, MarkView, ViewMutationRecord } from 'prosemirror-view'
 
-import { createKey } from '../create-key'
+import { createKey } from '../utils/create-key'
 import { isContentDOMRemoval } from '../utils/is-content-dom-removal'
 
 import type { CoreMarkViewSpec, CoreMarkViewUserOptions, MarkViewDOMSpec } from './CoreMarkViewOptions'
