+++ packages/core/src/nodeView/CoreNodeView.ts
@@ -2,7 +2,7 @@ import { isElementLike } from '@ocavue/utils'
 import type { Attrs, Node } from 'prosemirror-model'
 import type { Decoration, DecorationSource, EditorView, NodeView, ViewMutationRecord } from 'prosemirror-view'
 
-import { createKey } from '../create-key'
+import { createKey } from '../utils/create-key'
 import { isContentDOMRemoval } from '../utils/is-content-dom-removal'
 
 import type { CoreNodeViewSpec, CoreNodeViewUserOptions, NodeViewDOMSpec } from './CoreNodeViewOptions'
