+++ packages/core/src/pluginView/CorePluginView.ts
@@ -1,7 +1,7 @@
 import type { EditorState, PluginView } from 'prosemirror-state'
 import type { EditorView } from 'prosemirror-view'
 
-import { createKey } from '../create-key'
+import { createKey } from '../utils/create-key'
 
 import type { CorePluginViewSpec, CorePluginViewUserOptions } from './CorePluginViewOptions'
